Enable continuous streaming of eventhub data
